In Cosmos’ realm where modules meet, depinject guides with methods neat. Provide the lifeblood, functions share, Dependencies handled with care.
Supply the needs, no gaps in sight,
Inject the logic, pathways bright.
Config shapes the structured flow,
Design patterns help us grow.

In depinject, the future’s clear,
Dependency’s path, we hold dear.
